requirements.txt
Flask==3.1.0
gunicorn==23.0.0
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
cryptography>=43.0
Flask-Login==0.6.3
